v0.2.1
I decided not to change the database to mongodb (although it would most likely be a better fit here) since this would require changing too much, and I did not see such a strong need, so I created models for separate storage of the “final data” (which probably needed to be normalized, but we can only say for sure after speed tests).
An idempotent post-initialization command was added to populate models that store the necessary predictions.
Added calculation of lunar cycles.