v0.3.4
Now the service workers properly caching the necessary data and intercept requests. Nginx now returns nothing for Curl http://sitename/ and when switching to this address there will be a redirect to the most suitable language. This method is the best solution for a multi -language site, since it won't reduce SEO